+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.8-alpha-jammy
 
 * Remove package `update-manager-core` to avoid the user upgraded to Ubuntu next version. (AnduinOS will handle the upgrade)
+* Use totem as the default video player instead of vlc.
 
 ## v0.0.7-alpha-jammy